Added support for patrons, added patron list to api
Patrons are determined by the role given by the Patreon bot, once Lootcord detects the role added, it adds their ID to patron table. The patrons are automatically updated when the bot starts to keep the list current. An api endpoint for getting patrons along with their avatars has been added so that the website can retrieve patron information.