Fixed commandhandler false-positives
Sending a dm such as ok or ko caused the commandhandler to default to the backpack (I suspect because its the first js file in commands direc) command. In order to combat this, I had the handler also check if the commandName variable was null/undefined along with the command variable.