v4.0.1 Fixed issue where bot would randomly stop responding to commands
Issue was caused by the prefix query in app.js retrieving the wrong default prefix for every guild. Was caused because the prefix variable that set the default prefix to t- was outside of the scope of client.on('ready'). Resolved by adding an else statement that would change the prefix variable to t- if none was set.