v4.3.4 attempt to fix votes
If the user voted before their voteCounter reset, they would still have a voteTimeLeft variable set to a number. If the bot was restarted while the voteTimeLeft was still set, it would add them to both the voteTime setTimeout AND a voteTimeLeft setTimeout which was shorter than 12 hours, making it impossible to vote before their counter was reset. I solved this by resetting voteTimeLeft back to 0 every time the user votes.